Inclusão de validação para impedir que um registro receba uma nova edição antes da aprovação.

+++ dspace-api/src/main/java/org/dspace/content/InstallItemServiceImpl.java
@@ -88,9 +88,13 @@ public class InstallItemServiceImpl implements InstallItemService
         		if(CollectionUtils.isNotEmpty(valores))
             	{
             		Item oldItem = itemService.find(c, UUID.fromString(valores.get(0).getValue()));
-            		suppliedHandle = oldItem.getHandle();
-            	    
-            		itemService.delete(c, oldItem);
+            		
+            		if(oldItem != null)
+            		{
+            			suppliedHandle = oldItem.getHandle();
+            			itemService.delete(c, oldItem);
+            		}
+            				
             		handleService.removerItemDoHandle(c, item.getHandle());
             	}
     		}
